+++ vm_control/src/balloon_tube.rs
@@ -33,9 +33,9 @@ pub enum BalloonControlCommand {
     Stats,
     WorkingSet,
     WorkingSetConfig {
-        bins: Vec<u64>,
-        refresh_threshold: u64,
-        report_threshold: u64,
+        bins: Vec<u32>,
+        refresh_threshold: u32,
+        report_threshold: u32,
     },
 }
